Create Split-Sort-Tab
From an end-user's perspective, this is an Excel macro that checks the active sheet for semicolons (or whatever other character you assign to the "oldDivider" variable), splits any rows that include them into a corresponding number of rows with one entry each and otherwise identical data, and then creates one tab for each unique entry in the first column, populated by their matching rows.
Seemed potentially extensible enough to be useful to someone else. 
To do still: split the three functions into their own modules; add user input for which character to use as a delimiter; add user input for which column to split on.